+++ import_export/widgets.py
@@ -486,10 +486,12 @@ class JSONWidget(Widget):
         """
         :return: A JSON formatted string derived from ``value``.
           ``coerce_to_string`` has no effect on the return value.
+          Set the ``sort_json_keys`` keyword argument to ``True`` to sort the
+          keys in JSON objects.
         """
         if value is None:
             return None
-        return json.dumps(value, sort_keys=True)
+        return json.dumps(value, sort_keys=kwargs.get("sort_json_keys", False))
 
 
 class ForeignKeyWidget(Widget):